Add new Hugo site and update .gitignore
The commit creates new files for a Hugo site including HTML templates, CSS, a basic configuration file and sets up .gitignore for the project. The HTML files added are for pages such as "404 - Page Not Found", "Categories", "Tags", and the main homepage. The .gitignore files ensure editor and OS-specific files do not get committed into the repository.