Add post for Sundown Sessions #6
A new post has been created for the Sundown Sessions #6, which includes details about the content of the session, the playlist, and show notes. The featured image for the session has been added, and the post has been marked as a draft. Additionally, the post for Sundown Sessions #5 has been renamed.