Disable git configuration in Azure workflow
The changes made in the Azure static web-apps workflow in this commit have disabled the git configuration steps. This includes removal of actions that were previously checking out the repository and configuring git safety directory in the workflow. These steps were unnecessary for the Build And Deploy Job, hence removed.