Remove "Chapter II: Pierre Gringoire" post and update "Sundown Sessions" post
The "Chapter II: Pierre Gringoire" post was deleted from the site's content. In addition to this, modifications were made to the "Sundown Sessions Featuring Blue on Shock" post. These changes include adjusting the formatting of the "Playlist" header, and adding an additional section for "Show Notes", which includes "Additional Resources" and "Track Info".