Update site link and post details
This commit updates the link in src/public/index.xml and src/public/index.html from localhost to Github pages. It also changes the post date in src/content/posts/003-sundown-sessions-featuring-blue-on-shock.md and adjusts a wording in src/content/posts/_index.md. Most of the RSS feed contents are removed for clarity.