Add recent shows to site's index page
Updated index.html and index.xml files to include information about the recent shows on the site's main page. This is intended to increase the visibility of the newest content to site visitors. In addition, the Hugo generator version was updated in the meta tags of the index.html file.