Refactor show #1 content structure and update Hugo version
The main changes in this commit include restructuring the content for show #1 into separate markdown files, and updating the Hugo version being used. This gives a cleaner structure to the content and makes it easier to maintain. Additionally, updates have been made to the public index.html and sitemap.xml files reflecting these changes.