Update 'About' page content and associated metadata
The 'About' page content has been significantly updated and moved to a new location as part of this commit. The change reflects a shift in the focus of the site which is now centered towards serving the local community radio station. In addition, the associated metadata and site mappings have been updated to accurately reflect this change.