Remove French version of index and enhance index details
This commit removes the French version of the index file and improves the details on various index-related sections. Changes have been made to improve information clarity by making modifications to the description and renaming image files for better consistency. Lastly, timestamps in the sitemap have been updated to reflect the latest changes.