Update templates and modify sitemap timestamps
The changes include an update to the show and default templates used in the application. The show template order was altered, with the 'date' value moved for better readability. Single quotes were introduced to the 'featured_image' and 'read_more_copy' fields in the default template. Additionally, changes were also made to the timestamps in the