Add ID to form and implement AJAX submission
The contact form has been updated with an ID for better targeting. A JavaScript section has also been added which prevents the form from refreshing the page on submission and instead sends the form data via an AJAX fetch request to a specified URL. This will provide a more seamless user experience. In case of unsuccessful submission, errors are appropriately handled and logged #35.