Add wikilinks to artist names in the playlist
Wikilinks were added to all artist names in the playlist ensuring the names now link to new individual artist pages. These pages were also created for each artist and are populated with a title and featured image. This was facilitated by the creation of a new shortcode, artist-wikilink, to simplify the linking process.