Remove David Latto & The Lang Spoons and add new artist profiles
This commit removes the band profile for David Latto & The Lang Spoons and updates references to the artist. This includes updates to playlist and track info files. Additionally, several new artist profiles, including Del Shannon, The Detroit Cobras, and The Vermin Poets, have been added to the content directory. Minor changes have been made to existing artist files and to show logo in work-in-progress file for better cohesion.