Update path for featured images
The code has been revised to update the path construction for featured images in the "GetFeaturedImage.html" file. It now places them based on the first character of their name in the artists directory. The change also includes an update to the path of the featured image in "queens-of-the-stone-age.md".