Update show details and keywords in index.md
Several changes have been implemented in the show details and keywords for the 3rd session's index file. These changes include the addition of new keywords related to the show and featured guest, as well as an adjustment to the show's broadcast date. The featured guest's name has also been updated.