Update artist information across several pages
This commit includes updates to various artist pages. Changes include a new line indicating the death of Alan Rankine, updates to the current lineup of The Jags, a genre change for the band Cigarettes After Sex, the addition of the 'artist_page' metadata to The Big Nowhere and The Receiving End of Sirens, a brief detail addition to Graham Coxon's about section, among others.