Update and expand artist database
Several changes have been made to improve the accuracy and comprehensiveness of the artist database. This includes the correction of several misidentified artists or unclear references, as well as the addition of several new artist entries. In particular, detailed profiles of numerous artists have been included, enriching the overall content of the database.