Update track information for 'The Big Now'
This commit updates the details for two tracks by 'The Big Now'. The description is updated to specify that the tracks are from a demo cassette named 'Fast Cars, Soul Music' from the late 80s/early 90s. A note is added requesting information for the track '(Oooo La La) You're My Good Friend'.