Update featured guest link generation
This commit changes how the URL for the featured guest is generated. Instead of hardcoding the beginning of the URL (which was previously set to "http://localhost:1313/shows/featuring-the-big-now/#featured-guest-"), it now relies on .Site.BaseURL to form the base of the URL, making the URL generation more flexible and applicable to different environments. #117