Update show details and artist information
Numerous modifications have been made to the track list, playlist, and featured guest information for show 6. Additionally, artist details for "The Adventures," "Billie Eilish," and "a-ha" have been updated or added. The changes reflect the updated line-up of tracks and artists for the particular show, and aim to provide updated and accurate information for the viewers and listeners.